Fix dangling line calculated p & q
