+++ src/main/java/com/powsybl/openloadflow/network/impl/LfDanglingLineBranch.java
@@ -7,19 +7,26 @@
 package com.powsybl.openloadflow.network.impl;
 
 import com.powsybl.iidm.network.DanglingLine;
-import com.powsybl.openloadflow.network.AbstractFictitiousBranch;
+import com.powsybl.openloadflow.network.AbstractLfBranch;
 import com.powsybl.openloadflow.network.LfBus;
 import com.powsybl.openloadflow.network.PerUnit;
 import com.powsybl.openloadflow.network.PiModel;
+import com.powsybl.openloadflow.util.Evaluable;
 
 import java.util.Objects;
+
+import static com.powsybl.openloadflow.util.EvaluableConstants.NAN;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public class LfDanglingLineBranch extends AbstractFictitiousBranch {
+public class LfDanglingLineBranch extends AbstractLfBranch {
 
     private final DanglingLine danglingLine;
+
+    private Evaluable p = NAN;
+
+    private Evaluable q = NAN;
 
     protected LfDanglingLineBranch(DanglingLine danglingLine, LfBus bus1, LfBus bus2) {
         super(bus1, bus2, new PiModel(danglingLine.getR(), danglingLine.getX())
@@ -46,8 +53,28 @@ public class LfDanglingLineBranch extends AbstractFictitiousBranch {
     }
 
     @Override
+    public void setP1(Evaluable p1) {
+        // nothing to do
+    }
+
+    @Override
+    public void setP2(Evaluable p2) {
+        this.p = Objects.requireNonNull(p2);
+    }
+
+    @Override
+    public void setQ1(Evaluable q1) {
+        // nothing to do
+    }
+
+    @Override
+    public void setQ2(Evaluable q2) {
+        this.q = Objects.requireNonNull(q2);
+    }
+
+    @Override
     public void updateState() {
-        danglingLine.getTerminal().setP(p.eval() * PerUnit.SB);
-        danglingLine.getTerminal().setQ(q.eval() * PerUnit.SB);
+        danglingLine.getTerminal().setP(-p.eval() * PerUnit.SB);
+        danglingLine.getTerminal().setQ(-q.eval() * PerUnit.SB);
     }
 }
